+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added support for `SourceDescription.poster` for Android & iOS.
 - Made the intervals for the forward and backward skip buttons on the iOS lockScreen configurable.
+- Added preferredPeakBitRate and preferredMaximumResolution to ABRConfiguration for iOS
 
 ## [7.3.0] - 24-06-03